third_party/blink/web_tests/flag-specific/highdpi/virtual/container-queries/wpt_internal/css/css-conditional/container-queries/at-container-006-expected.txt

This is a testharness.js-based test.
FAIL Pseudo-element ::backdrop responds to container size changes assert_equals: expected "rgb(0, 0, 0)" but got "rgb(0, 128, 0)"
Harness: the test ran to completion.
